requirements.txt
langgraph == 0.2.66
langsmith == 0.2.11
langchain-text-splitters  == 0.3.5
langchain-community == 0.3.15
langchain-pinecone == 0.2.2
langgraph-checkpoint == 2.0.10
langchain-openai == 0.3.2

jq == 1.8.0
pinecone == 5.4.2
pypdf == 5.1.0

python-dotenv == 1.0.1

PyMuPDF == 1.25.2
arxiv == 2.1.3
